+++ shop_service/src/main/java/com/luckytree/shop_service/shop/adapter/data/ShopDetail.java
@@ -2,7 +2,7 @@ package com.luckytree.shop_service.shop.adapter.data;
 
 import com.luckytree.shop_service.common.enums.Category;
 import com.luckytree.shop_service.common.enums.Hashtag;
-import com.luckytree.shop_service.shop.adapter.persistence.ShopEntity;
+import com.luckytree.shop_service.shop.adapter.jpa.ShopEntity;
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.*;
