+++ shop_service/src/main/java/com/luckytree/shop_service/shop/adapter/data/ShopDetailDto.java
@@ -2,7 +2,7 @@ package com.luckytree.shop_service.shop.adapter.data;
 
 import lombok.*;
 import com.luckytree.shop_service.common.enums.Category;
-import com.luckytree.shop_service.shop.adapter.persistence.ShopEntity;
+import com.luckytree.shop_service.shop.adapter.jpa.ShopEntity;
 import com.luckytree.shop_service.common.enums.Hashtag;
 
 @Getter
