+++ shop_service/src/main/java/com/luckytree/shop_service/shop/adapter/jpa/ShopRemoveRepository.java
@@ -1,4 +1,4 @@
-package com.luckytree.shop_service.shop.adapter.persistence;
+package com.luckytree.shop_service.shop.adapter.jpa;
 
 import org.springframework.data.jpa.repository.JpaRepository;
