+++ shop_service/src/main/java/com/luckytree/shop_service/shop/adapter/jpa/ShopRepository.java
@@ -1,4 +1,4 @@
-package com.luckytree.shop_service.shop.adapter.persistence;
+package com.luckytree.shop_service.shop.adapter.jpa;
 
 import com.luckytree.shop_service.common.enums.Category;
 import com.luckytree.shop_service.common.enums.Hashtag;
