+++ shop_service/src/main/java/com/luckytree/shop_service/shop/application/port/outgoing/ShopPort.java
@@ -3,7 +3,7 @@ package com.luckytree.shop_service.shop.application.port.outgoing;
 import com.luckytree.shop_service.shop.adapter.data.BookmarkDto;
 import com.luckytree.shop_service.common.enums.Category;
 import com.luckytree.shop_service.shop.adapter.data.ShopRequest;
-import com.luckytree.shop_service.shop.adapter.persistence.ShopEntity;
+import com.luckytree.shop_service.shop.adapter.jpa.ShopEntity;
 import com.luckytree.shop_service.common.enums.Hashtag;
 import com.luckytree.shop_service.shop.adapter.data.ShopDetail;
 import com.luckytree.shop_service.shop.adapter.data.ShopSummary;
